Support dict type for CopySource
Instead of trying to parse the CopySource value and
encoded values for the customer, they can instead provide
a Bucket, Key, and optional VersionId dict for the value
of CopySource.  This removes all the previous edge cases,
and removes the requirement that the customer know the header
serialization format for copy source.

Related to #757.