Remove caching in request signer
This simplifies the code and fixes potential edge cases from
caching.

This was a legacy artifact from having to support both the
service/operation objects and the client interfaces at the
same time.  This is now no longer needed.

The only concern would be if caching actually helps performance.
To verify this, I ran a test consisting of:

* Mocking out the _send_request in endpoints.py to return a precanned
  response.
* Making a client call 1000 times (I chose ec2 DescribeInstances
  arbitrarily).  This is essentially timing the time it takes from
  a client method until it hits the http layer in botocore.

I ran the tests against botocore's develop branch and this PR.  There's
negligible difference::

    (simple-auth %) $ repeat 10 python scripts/signerperf.py
    Total time: 0.7235
    Total time: 0.7022
    Total time: 0.7190
    Total time: 0.7032
    Total time: 0.7112
    Total time: 0.7059
    Total time: 0.7099
    Total time: 0.6895
    Total time: 0.7079
    Total time: 0.7254
    (simple-auth %) $ git checkout develop
    Switched to branch 'develop'
    (develop %) $ repeat 10 python scripts/signerperf.py
    Total time: 0.7085
    Total time: 0.6812
    Total time: 0.7016
    Total time: 0.6995
    Total time: 0.7055
    Total time: 0.7314
    Total time: 0.6950
    Total time: 0.7105
    Total time: 0.7140
    Total time: 0.7129