Don't preemptively convert unicode in rest-xml parser
In the rest xml parser, we were converting the input we recieve into a
`str`, which was causing failures on python 2 when multibyte unicode
strings were passed in. The fix is to simply use `six.text_type`, which
is `unicode` on 2 and `str` on 3. The string will be encoded into the
default encoding later on in the serializer.

Fixes #868