Automatically redirect S3 sigv4 requests sent to the wrong region
S3 generally provides enough information for us to redirect
requests, so this attempts to do so. To prevent a massive
number of additional requests, bucket regions are cached.

When a redirect occurs, a warning is printed that tells the
customer they should use a client configured to the proper
region to avoid additional requests.