Add support for use_dualstack_endpoint in clients
I wanted to make this change in as logical a manner as
possible without having to touch every corresponding piece
of code that needed new behavior.  There were a few things
that needed to be handled:

* Handling the proper DNS suffix for non aws partitions (cn-north-1)
* Working together with the "addressing_style" config.
* Working with the S3 x-amz-region redirect code.  If we're in dualstack
  mode, we should still use dualstack endpoints when following
  x-amz-region

Dualstack support is implemented by updating the
ClientEndpointBridge to be "dual stack aware".  The rationale
for this is that any code that is using the client bridge
will automatically get this behavior without any code changes.

On a practical level, this means that the x-amz-region redirect
code will "just work" because it's relying on the endpoint bridge
to do all the heavy lifting.

This also means that the logic for virtual vs. path style mostly
just works, with one caveat.  The fix_s3_host previously had a
hardcoded "s3.amazonaws.com".  When dualstack is requested, we need
to provide a default_endpoint_url of None, which means that it will
use the existing request url to perform the virtual host switching
logic.  Because the client bridge has already constructed the
dual stack endpoint, fix_s3_host will now just work.