Remove a bit of dead code from s3 presign default
There was a bit of code that checked to make sure that the service
was actually S3 and that the signature version wasn't unsigned.
However, there will never be a case where either of those conditions
will be met because the handler just won't be registered in the first
place. Since the code never gets called, this removes it.