Discard starting_truncation after first request
This fixes a bug where we were not discarding the starting truncation
value in the paginator after the first request was made and truncated.
This resulted in having an inflated `boto_truncate_amount` value in
the pagination token. Successive calls would begin to drop results
starting at the third call.