Fix mutation issue in building retry config
Even prior to this feature, when a retry config was built
if a service had specific configurations it would mutate
the cached retry model. So for example if a user created
a dynamodb client and then an s3 client, the s3 client would
inherit the max attempts of 10 from the dynamodb client.

Tests were also added for checking that the max attempts
provided to the session can be overriden at the client
level.