Fix s3 addressing for unsigned requests in cn-north-1
An unsigned request would default to the existing host of
's3.amazonaws.com' when attempting to switch to virtual
host addressing.  This is incorrect for regions outside of
the 'aws' partition.  The fix here is to use a default
endpoint url of `None` for `fix_s3_host`, which will result
in parsing out the hostname from the url when constructing
the virtual hosted address.