Fix PEP 479 test issues
PEP 479 changed the way StopIteration affects generators. Unfortunately
a couple of our tests were silenty and unintentionally relying on the
previous behavior because Mock will raise StopIteration if you have
given it a list of side effects and that list is empty when it goes to
fetch the next response.