Improve error message with missing service id
Originally introduced in 9250fb231a35472788b6db0c444202f544a42f72,
one of the tests was modified to have different behavior than intended.

The test is suppose to verify that without any of the needed
metadata properties we don't register a retry handler.
Right now, if a serviceId is missing you'll get an error message of
"NoneType" has no attribute "lower()".
With this change, you'll get the missing metadata error message
from models.py.

To help clarify this, I've changed the name of the test.  I also
added a repr to the Service class.  You now get:

$ aws s3 ls

"serviceId" not defined in the metadata of the model: ServiceModel(s3)