Add property methods support for get_config_variable in session
If a config variable's loading precendnce was configured  using
old-style configuration override tuples provided to the botocore
session.session_vars_map, it would then support the methods argument
to get_config_variable correctly. However; if it was configured using
the new style providers, it would not. This adds another small
backwards compatibility layer to add support for the methods argument
in the case of the default botocore configuration logical names.