Support providing access-point ARNs to S3 APIs
If an access-point arn is provided to the Bucket parameter, the
arn will be parsed and endpoint will be created using information
from the arn representing the access-point's endpoint url. Added
configuration options as well to chose between using an ARN's
region and a client's region when constructing the endpoint.