JSONFileLoader: dedent unnecessary code inside context manager
In theory, there should be no problem parsing the JSON inside the context
manager (while the file handle is open), it somehow allowed a bug to be
triggered whereby the file descriptor would be closed prematurely. This
caused an `OSError` to be raised ("[Errno 9] Bad file descriptor") when
being used in a multiprocessing environment.

The "fix" here is simply to do less work inside the context manager.
I have found that this prevents the bug from being triggered, but can't
explain exactly why!